feat(groupGenerator): 将逻辑整合到一个 Generator 里面，用一个开关控制
